storage mock fix
