+++ storage/__mocks__/pm2.js
@@ -1,0 +1,11 @@
+module.exports = {
+    list: (callback) => {
+        callback(null, [{
+            name: `motion`,
+            pm2_env: {
+                status: "on",
+                unstable_restarts: 0
+            }
+        }])
+    }
+}
