+++ storage/test/convert.test.js
@@ -1,10 +1,7 @@
 const supertest = require('supertest');
 const app = require('../backend/storage.js')
-const command = require('command').app
-const path       = require("path")
-var {generateID, fileName, parseFileName}    = require("../backend/routes/lib/converter.js")
 
-const imgDir = path.join(process.env.storage_FILEPATH, "shared/captures")
+var {generateID, fileName, parseFileName}    = require("../backend/routes/lib/converter.js")
 
 const fileList = [
     "output_1_20210101-235959_20210201-235959_video1-20210301-235959.mp4", 
@@ -21,81 +18,13 @@ const parsedFileList = fileList.filter(file => !file.includes('txt')).map((file)
     }
 })
 
-const { testLib } = require('lib')
-const { mockedPassword, hashedMockedPassword } = testLib
-let fs = require('fs')
-fs.readFile = jest.fn().mockImplementation((filePath, options, callback) => {
-    if(options instanceof Function){
-        options(false, {})
-    }
-    else{
-        callback(false, hashedMockedPassword)
-    }
-})
-fs.readdir = jest.fn().mockImplementation((filePath, callback) => {
-    if(filePath == imgDir){
-        callback(false, fileList)
-    }
-    else{
-        callback(true, [])
-    }
-})
-fs.stat = jest.fn().mockImplementation((filePath, callback) => {
-    if(fileList.map(file => path.join(imgDir, file)).includes(filePath)){
-        callback(false)
-    }
-    else{
-        callback(true)
-    }
-})
-fs.unlink = jest.fn().mockImplementation((filePath, callback) => {
-    if(fileList.map(file => path.join(imgDir, file)).includes(filePath)){
-        callback(false)
-    }
-    else{
-        callback(true)
-    }
-})
-
-jest.mock('memory', () => ({
-    client: (name) => ({
-        emit: (event, ...args) => {
-            if(event == "savePassword"){
-                args[1]()
-            }
-            else if(event == "verifyPassword"){
-                args[1](false)
-            }
-        },
-        on: () => {}
-    }),
-    server: () => {}
-}))
-
-jest.mock('pm2', () => ({
-    list: (callback) => {
-        callback(null, [{
-            name: `motion`,
-            pm2_env: {
-                status: "on",
-                unstable_restarts: 0
-            }
-        }])
-    }
-}))
+jest.mock('lib')
+jest.mock('fs')
+jest.mock('memory')
+jest.mock('pm2')
 
 describe('Convert Routes', () => {
-    let cookieWithBearerToken
-    beforeAll((done) => {
-        supertest(command)
-        .post('/authorization/login')
-        .send({password: mockedPassword})
-        .expect(200)
-        .expect('set-cookie', /bearertoken=Bearer%20.*; Max-Age=.*/, (err, res) => { 
-            cookieWithBearerToken = res.headers["set-cookie"]
-            done()
-        })
-    })
+    let cookieWithBearerToken = "validCookie"
 
     describe("/convert/createVideo", () => {
         test('bruh', () => expect(2+2).toBe(4))
